Add default value (0) to clap migration

+++ database/migrations/2019_05_08_080657_add_clap_column_to_answers.php
@@ -14,7 +14,7 @@ class AddClapColumnToAnswers extends Migration
     public function up()
     {
         Schema::table('answers', function (Blueprint $table) {
-            $table->integer('claps')->default(0);
+            $table->bigInteger('claps')->default(0);
         });
     }
